logs sending to front end

+++ app/data_classes/message_result_modal.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 @dataclass
 class Predictions():
-    confidence: int
+    confidence: float
     violation: str
     violation_id: int
     detection_id: int
